# Switch Murmur3 to faster implementation