--- a/go.mod
+++ b/go.mod
@@ -5,8 +5,8 @@ go 1.21
 toolchain go1.21.1
 
 require (
-	github.com/spaolacci/murmur3 v1.1.0
 	github.com/stretchr/testify v1.8.4
+	github.com/twmb/murmur3 v1.1.8
 )
 
 require (