--- a/thetacommon/theta_utils.go
+++ b/thetacommon/theta_utils.go
@@ -18,5 +18,5 @@
 package thetacommon
 
 const (
-	DefaultUpdateSeed = uint32(9001)
+	DefaultUpdateSeed = uint64(9001)
 )