Switch Murmur3 dependency to https://github.com/twmb/murmur3 as it yield better performance than the current one.

```
pkg: github.com/apache/datasketches-go/hll
                                  │      before.bench      │             after.bench             │
                                  │         sec/op         │   sec/op     vs base                │
HLLDataSketch/lgK4_HLL4_uint-10               13.365n ± 1%   9.237n ± 2%  -30.88% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL4_uint-10              13.535n ± 1%   9.496n ± 1%  -29.84% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL4_uint-10               18.96n ± 1%   13.04n ± 1%  -31.25% (p=0.000 n=10)
HLLDataSketch/lgK4_HLL6_uint-10               13.545n ± 1%   9.361n ± 0%  -30.89% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL6_uint-10              13.560n ± 1%   9.504n ± 1%  -29.91% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL6_uint-10               18.55n ± 0%   12.76n ± 1%  -31.23% (p=0.000 n=10)
HLLDataSketch/lgK4_HLL8_uint-10               12.595n ± 1%   8.440n ± 1%  -32.99% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL8_uint-10              12.670n ± 1%   8.460n ± 1%  -33.23% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL8_uint-10               17.34n ± 1%   11.28n ± 1%  -34.97% (p=0.000 n=10)
HLLDataSketch/lgK4_HLL4_slice-10              13.445n ± 0%   9.716n ± 1%  -27.74% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL4_slice-10             13.670n ± 1%   9.731n ± 1%  -28.82% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL4_slice-10              19.09n ± 0%   13.32n ± 3%  -30.21% (p=0.000 n=10)
HLLDataSketch/lgK4_HLL6_slice-10              13.570n ± 0%   9.803n ± 1%  -27.76% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL6_slice-10             13.770n ± 1%   9.826n ± 0%  -28.64% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL6_slice-10              18.85n ± 1%   12.81n ± 1%  -32.06% (p=0.000 n=10)
HLLDataSketch/lgK4_HLL8_slice-10              12.765n ± 1%   8.835n ± 8%  -30.79% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL8_slice-10             12.880n ± 1%   8.870n ± 1%  -31.13% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL8_slice-10              17.69n ± 1%   11.54n ± 1%  -34.72% (p=0.000 n=10)
HLLDataSketch/lgK4_HLL8_union-10               26.76n ± 1%   22.68n ± 1%  -15.21% (p=0.000 n=10)
HLLDataSketch/lgK16_HLL8_union-10              26.60n ± 1%   22.68n ± 1%  -14.76% (p=0.000 n=10)
HLLDataSketch/lgK21_HLL8_union-10              26.50n ± 1%   22.68n ± 0%  -14.43% (p=0.000 n=10)
geomean                                        16.10n        11.45n       -28.87%
```