@AlexanderSaydakov 2023-12-19
Have you tried to test compatibility with C++ and Java? Even a slightest difference in hash function can make sketches incompatible (disjoint sets despite of the same input)