@freakyzoidberg 2023-12-19
Checked a few hash output for equality.

I am actually thinking of checking back in the serialised output test data, to test for possible change like this.


Anyhow, in this case, crc32 for the various generations match with both implementations

```
❯ sdiff  <(crc32 ./go_generated_files/*) <(crc32 ./go_generated_files_old/*)
422a5893	./go_generated_files/hll4_n0_go.sk            | 422a5893	./go_generated_files_old/hll4_n0_go.sk
dcb6ae94	./go_generated_files/hll4_n1000000_go.sk      | dcb6ae94	./go_generated_files_old/hll4_n1000000_go.sk
415450fd	./go_generated_files/hll4_n100000_go.sk       | 415450fd	./go_generated_files_old/hll4_n100000_go.sk
9942a074	./go_generated_files/hll4_n10000_go.sk        | 9942a074	./go_generated_files_old/hll4_n10000_go.sk
44466b15	./go_generated_files/hll4_n1000_go.sk         | 44466b15	./go_generated_files_old/hll4_n1000_go.sk
cee02cc2	./go_generated_files/hll4_n100_go.sk          | cee02cc2	./go_generated_files_old/hll4_n100_go.sk
f6ac7c4e	./go_generated_files/hll4_n10_go.sk           | f6ac7c4e	./go_generated_files_old/hll4_n10_go.sk
f39fce1b	./go_generated_files/hll4_n1_go.sk            | f39fce1b	./go_generated_files_old/hll4_n1_go.sk
45479c8a	./go_generated_files/hll6_n0_go.sk            | 45479c8a	./go_generated_files_old/hll6_n0_go.sk
8f48e75f	./go_generated_files/hll6_n1000000_go.sk      | 8f48e75f	./go_generated_files_old/hll6_n1000000_go.sk
cf60fc77	./go_generated_files/hll6_n100000_go.sk       | cf60fc77	./go_generated_files_old/hll6_n100000_go.sk
2a06d1cb	./go_generated_files/hll6_n10000_go.sk        | 2a06d1cb	./go_generated_files_old/hll6_n10000_go.sk
96f9cd30	./go_generated_files/hll6_n1000_go.sk         | 96f9cd30	./go_generated_files_old/hll6_n1000_go.sk
b37ef4b7	./go_generated_files/hll6_n100_go.sk          | b37ef4b7	./go_generated_files_old/hll6_n100_go.sk
1567d7bd	./go_generated_files/hll6_n10_go.sk           | 1567d7bd	./go_generated_files_old/hll6_n10_go.sk
061f68db	./go_generated_files/hll6_n1_go.sk            | 061f68db	./go_generated_files_old/hll6_n1_go.sk
4cf1d0a1	./go_generated_files/hll8_n0_go.sk            | 4cf1d0a1	./go_generated_files_old/hll8_n0_go.sk
151be68f	./go_generated_files/hll8_n1000000_go.sk      | 151be68f	./go_generated_files_old/hll8_n1000000_go.sk
c1170afe	./go_generated_files/hll8_n100000_go.sk       | c1170afe	./go_generated_files_old/hll8_n100000_go.sk
3c8bcc28	./go_generated_files/hll8_n10000_go.sk        | 3c8bcc28	./go_generated_files_old/hll8_n10000_go.sk
935c33e5	./go_generated_files/hll8_n1000_go.sk         | 935c33e5	./go_generated_files_old/hll8_n1000_go.sk
35dd9c28	./go_generated_files/hll8_n100_go.sk          | 35dd9c28	./go_generated_files_old/hll8_n100_go.sk
ea4a2de9	./go_generated_files/hll8_n10_go.sk           | ea4a2de9	./go_generated_files_old/hll8_n10_go.sk
c3ef85da	./go_generated_files/hll8_n1_go.sk            | c3ef85da	./go_generated_files_old/hll8_n1_go.sk
```